requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
openai>=1.3.7
qdrant-client>=1.8.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.5.2
python-multipart>=0.0.6
sentence-transformers>=2.2.2